Revert "First try to integrate axp209 into nand uboot code"
This reverts commit 1b3ebd9524792b613a3995e78be9de836ff52ae8.

I don't think we need the axp209 code inside our u-boot.  This was originally
worked on trying to add the power control capability to tackle power issues, but
later we discovered that is not the case.

We ended up getting source code of AW boot0/boot1 from Olimex's Android SDK,
building the boot0/boot1 binaries from there and used the customized binaries.